Add HTTPD support
The use case for this feature is so scripts that utilize Hubot's router
(HTTPD) features can be tested. I (and many others) have several Hubot
scripts that respond to GET and POST requests via the Hubot's Web Hook
API. I use these to configure SSH keys stored in Redis (robot.brain) and
for sending notifications from build/deploy scripts so Hubot can
announce to users when a build is complete.