feat: Update hubot dependency to allow for hubot 3
There has been a fair amount of work upstream in hubot recently. We are
really close to a 3.0 release (there is actually a `next` release of
it), and I wanted to make sure this library supported it since it is
recommended in the hubot documentation.

See https://github.com/hubotio/hubot/issues/1057 on this